# server/docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.9
paths:
...
  /mini/pay/create:
    post:
      consumes:
      - application/json
      description: 根据订单类型与订单 ID 生成预支付单（微信 V3），返回小程序调起支付所需参数（signType 为 RSA）。需登录，请求头必带
        x-token；须已完整配置微信支付，否则返回错误。
      parameters:
      - description: 小程序登录后返回的 token（必填）
        in: header
        name: x-token
        required: true
        type: string
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: data 含 timeStamp,nonceStr,package,signType,paySign
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
                msg:
                  type: string
              type: object
      summary: 调起支付
      tags:
      - 小程序
  /mini/pay/refund:
    post:
      consumes:
      - application/json
      description: 对已支付且待核销、且尚未产生核销次数的门票订单申请全额退款（多次票若已核销过任一次则不可退）。限时活动订单不支持用户自助退款，请联系客服。需登录，请求头必带
        x-token。
      parameters:
      - description: 小程序登录后返回的 token（必填）
        in: header
        name: x-token
        required: true
        type: string
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 申请退款
      tags:
      - 小程序
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...